# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Utility Management System API",
        "description": "API documentation for the Laravel Utility Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get a list of invoices",
                "description": "Returns a paginated list of all invoices.",
                "operationId": "9944e4d0532fc102a8e44d9f86025314",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoices retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/InvoiceResource"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 4
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoices not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get a single invoice",
                "description": "Returns the details of a specific invoice by its ID.",
                "operationId": "d02369d1b2059ace1cd3a9a67ffe2f30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the invoice",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Update an existing invoice",
                "description": "Updates the status or bill ID of an existing invoice.",
                "operationId": "692a6ee6cd992b205108bf6461d07d02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the invoice to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "billId": {
                                        "description": "ID of the associated bill",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Pending",
                                            "Paid",
                                            "Overdue"
                                        ],
                                        "example": "Paid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Invoice not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InvoiceResource": {
                "title": "Invoice Resource",
                "description": "Invoice model representation",
                "properties": {
                    "id": {
                        "description": "Invoice ID",
                        "type": "integer",
                        "example": 1
                    },
                    "billId": {
                        "description": "ID of the associated bill",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "Status of the invoice",
                        "type": "string",
                        "example": "Paid"
                    },
                    "createdAt": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-28T12:00:00.000000Z"
                    },
                    "updatedAt": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-28T12:00:00.000000Z"
                    },
                    "bill": {
                        "description": "Associated bill information",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "room_id": {
                                "type": "string",
                                "format": "uuid",
                                "example": "9a7c6f2c-5b8a-4f2a-8f2c-6d8b3a0c1e9f"
                            },
                            "user_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "rental_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 500
                            },
                            "electricity_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 50
                            },
                            "water_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 25
                            },
                            "fine_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 0
                            },
                            "service_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 10
                            },
                            "ground_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 5
                            },
                            "car_parking_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 15
                            },
                            "wifi_fee": {
                                "type": "number",
                                "format": "float",
                                "example": 20
                            },
                            "total_amount": {
                                "type": "number",
                                "format": "float",
                                "example": 625
                            },
                            "due_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2023-11-28"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-10-28T12:00:00.000000Z"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-10-28T12:00:00.000000Z"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}